Amanha é o niver da malu EBAAAAAA
Finalização das funcionalidades da tela de configurações: Alturalizar o cadastro, enviar uma avaliação e a exibição das avaliações

+++ ProjetoPratica_API/Models/Avaliacoes.cs
@@ -5,7 +5,7 @@ namespace ProjetoPratica_API.Data
     public class Avaliacoes
     {
         public int Id { get; set; }
-        public int QtdEstrelas { get; set; }
+        public byte QtdEstrelas { get; set; }
         public string Comentario { get; set; }
         public DateTime Data { get; set; }
         public int IdUsuario { get; set; }
